+++ js/data.js
@@ -1,0 +1,69 @@
+const avatar = [
+  'img/avatars/user01.png',
+  'img/avatars/user02.png',
+  'img/avatars/user03.png',
+  'img/avatars/user04.png',
+  'img/avatars/user05.png',
+  'img/avatars/user06.png',
+  'img/avatars/user07.png',
+  'img/avatars/user08.png',
+]
+
+const title = 'Студия 19кв.'
+
+const price = [
+  '11900',
+  '13500',
+  '15000',
+]
+
+const type = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+]
+
+const rooms = [
+  '1',
+  '2',
+  '3',
+  '4',
+]
+
+const guests = [
+  '5',
+  '7',
+  '9',
+]
+
+const checkin = [
+  '12:00',
+  '13:00',
+  '14:00',
+]
+
+const checkout = [
+  '12:00',
+  '13:00',
+  '14:00',
+]
+
+const features = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+]
+
+const description = 'описание помещения.'
+
+const photos = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+]
+
+export {avatar, title, price, type, rooms, guests, checkin, checkout, features, description, photos};
