+++ js/main.js
@@ -1,115 +1,5 @@
-//целое число из переданного диапазона включительно
-const getRandomIntInclusive = function(min, max) {
-
-  if(!isNaN(min) && !isNaN(max)) {
-
-    min = Math.ceil(min);
-    max = Math.floor(max);
-
-    if (min >= 0 && max >= 0) {
-      return (max > min) ? Math.floor(Math.random() * (max - min + 1)) + min : min;
-    }
-    return null;
-  }
-  return null;
-}
-
-//число с плавающей точкой из переданного диапазона включительно
-const getRandomDecimalInclusive = function(min, max, decimalPlaces) {
-
-  if(!isNaN(min) && !isNaN(max) && !isNaN(decimalPlaces)) {
-
-    if (min >= 0 && max >= 0) {
-      let random = (max > min) ? (min + Math.random() * (max - min)).toFixed(decimalPlaces) : min;
-      return parseFloat(random);
-    }
-    return null;
-  }
-  return null;
-}
-
-const avatar = [
-  'img/avatars/user01.png',
-  'img/avatars/user02.png',
-  'img/avatars/user03.png',
-  'img/avatars/user04.png',
-  'img/avatars/user05.png',
-  'img/avatars/user06.png',
-  'img/avatars/user07.png',
-  'img/avatars/user08.png',
-]
-
-const title = 'Студия 19кв.'
-
-const price = [
-  '11900',
-  '13500',
-  '15000',
-]
-
-const type = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-]
-
-const rooms = [
-  '1',
-  '2',
-  '3',
-  '4',
-]
-
-const guests = [
-  '5',
-  '7',
-  '9',
-]
-
-const checkin = [
-  '12:00',
-  '13:00',
-  '14:00',
-]
-
-const checkout = [
-  '12:00',
-  '13:00',
-  '14:00',
-]
-
-const features = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-]
-
-const description = 'описание помещения.'
-
-const photos = [
-  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
-]
-
-const getRandomArrayElement = (elemets) => {
-  return elemets[getRandomIntInclusive(0 , elemets.length - 1)];
-};
-
-//Набираем в массив случайное количество элементов
-const getRandomArrayIndicators = (elemets) => {
-  let randomArry = [];
-  for( let i = 0; i < elemets.length -1; i += 1) {
-    if (getRandomIntInclusive(0, 1) == true) {
-      randomArry.push(elemets[i]);
-    }
-  }
-  return randomArry;
-};
+import {avatar, title, price, type, rooms, guests, checkin, checkout, features, description, photos} from './data.js';
+import {getRandomIntInclusive, getRandomDecimalInclusive, getRandomArrayElement, getRandomArrayIndicators} from './utils.js'
 
 const apartmentDeals = () => {
   return [
@@ -142,4 +32,4 @@ const apartmentDeals = () => {
 
 const similarDeals = new Array(4).fill(null).map(() => apartmentDeals());
 
-similarDeals; // чтобы небыло ошибок в файле
+console.log(similarDeals); // для проверки
